Increase tempo to two actions per second.
It's doing about 15 hands per minute, and the django container's CPU is hovering around 60%.  Not bad!

    SELECT
      EXTRACT(
        EPOCH
        FROM
          MAX(CREATED) - MIN(CREATED)
      ) / COUNT(*)
    FROM
      PUBLIC.APP_HAND;

gives me a bit over two -- that's two seconds *per hand*.  Not bad!

+++ project/app/management/commands/big_bot_stress.py
@@ -22,7 +22,7 @@ everybodys_password = (
 
 class Command(BaseCommand):
     def add_arguments(self, parser) -> None:
-        parser.add_argument("--tempo-seconds", type=int, default=5)
+        parser.add_argument("--tempo-seconds", type=float, default=5.0)
 
         group = parser.add_mutually_exclusive_group()
         group.add_argument("--min-players", type=int, default=0)
